requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==3.0.0
attrs==24.2.0
audioread==3.0.1
blinker==1.9.0
cachelib==0.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dash==2.18.2
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-extensions==1.0.19
dash-html-components==2.0.0
dash-table==5.0.0
dash-uploader==0.6.1
dataclass-wizard==0.30.1
decorator==5.1.1
deepspeech==0.9.3
distro==1.9.0
EditorConfig==0.12.4
exceptiongroup==1.2.2
executing==2.1.0
ffmpeg-python==0.2.0
filelock==3.16.1
flask==3.0.3
Flask-Caching==2.3.0
fonttools==4.55.0
fpdf==1.7.2
fsspec==2024.10.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
ipython==8.18.1
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsbeautifier==1.15.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab-widgets==3.0.13
kiwisolver==1.4.7
lazy-loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
moviepy==2.1.1
mpmath==1.3.0
msgpack==1.1.0
narwhals==1.15.2
nest-asyncio==1.6.0
networkx==3.2.1
noisereduce==3.0.3
numba==0.60.0
numpy==2.0.2
openai==1.55.3
openai-whisper==20240930
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
pooch==1.8.2
proglog==0.1.10
prompt-toolkit==3.0.48
protobuf==5.29.1
psutil==6.1.0
pure-eval==0.2.3
pyarrow==18.1.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.2
pydantic-core==2.27.1
pydeck==0.9.1
pydub==0.25.1
pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rich==13.9.4
rpds-py==0.22.3
scikit-learn==1.5.2
scipy==1.13.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
soxr==0.5.0.post1
SpeechRecognition==3.11.0
stack-data==0.6.3
streamlit==1.40.2
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
toml==0.10.2
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
wcwidth==0.2.13
werkzeug==3.0.6
widgetsnbextension==4.0.13
zipp==3.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==3.0.0
attrs==24.2.0
audioread==3.0.1
bidict==0.23.1
blinker==1.9.0
cachelib==0.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dash==2.18.2
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-extensions==1.0.19
dash-html-components==2.0.0
dash-table==5.0.0
dash-uploader==0.6.1
dataclass-wizard==0.30.1
decorator==5.1.1
deepspeech==0.9.3
distro==1.9.0
EditorConfig==0.12.4
exceptiongroup==1.2.2
executing==2.1.0
ffmpeg-python==0.2.0
filelock==3.16.1
flask==3.0.3
Flask-Caching==2.3.0
Flask-SocketIO==5.4.1
fonttools==4.55.0
fpdf==1.7.2
fsspec==2024.10.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
ipython==8.18.1
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsbeautifier==1.15.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab-widgets==3.0.13
kiwisolver==1.4.7
lazy-loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
moviepy==2.1.1
mpmath==1.3.0
msgpack==1.1.0
narwhals==1.15.2
nest-asyncio==1.6.0
networkx==3.2.1
noisereduce==3.0.3
numba==0.60.0
numpy==2.0.2
openai==1.55.3
openai-whisper==20240930
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
pooch==1.8.2
proglog==0.1.10
prompt-toolkit==3.0.48
protobuf==5.29.1
psutil==6.1.0
pure-eval==0.2.3
pyarrow==18.1.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.2
pydantic-core==2.27.1
pydeck==0.9.1
pydub==0.25.1
pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rich==13.9.4
rpds-py==0.22.3
scikit-learn==1.5.2
scipy==1.13.1
simple-websocket==1.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
soxr==0.5.0.post1
SpeechRecognition==3.11.0
stack-data==0.6.3
streamlit==1.40.2
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
toml==0.10.2
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
wcwidth==0.2.13
werkzeug==3.0.6
widgetsnbextension==4.0.13
wsproto==1.2.0
zipp==3.21.0
